Add test for recovery phrase
